backend/requirements.txt
alembic==1.16.5
bcrypt==4.3.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
Flask-Cors==4.0.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.1
mirakuru==2.6.1
packaging==25.0
pluggy==1.6.0
port-for==0.7.4
psutil==7.1.0
psycopg==3.2.10
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-postgresql==7.0.2
python-magic==0.4.27
SQLAlchemy==2.0.43
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
WTForms==3.2.1
